requirements.txt
asgiref==3.4.1
Django==3.2.6
django-admin-thumbnails==0.2.6
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-environ==0.8.1
django-js-asset==1.2.2
django-modeltranslation==0.17.3
django-mptt==0.13.2
docopt==0.6.2
environs==9.3.3
generics==3.4.0
gitignore==0.0.8
gunicorn==20.1.0
marshmallow==3.13.0
Pillow==8.3.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
python-gettext==4.0
pytz==2021.1
six==1.16.0
slug==2.0
sqlparse==0.4.1
